Simplify the hell out of things.
* No more `element`, just use `byte`.
* No more `polynomial`, just use `[]byte`.
* Inline `add` and `sub`.
* Remove `log`. Use the table instead.
* Ditch test helpers. Literally no memory of what I was trying to do there.
* Use `byte` and `map[byte][]byte` for the public API.
* Re-flow some comments.

+++ sss_test.go
@@ -6,8 +6,8 @@ import (
 
 func Example() {
 	secret := "well hello there!" // our secret
-	n := 30                       // create 30 shares
-	k := 2                        // require 2 of them to combine
+	n := byte(30)                 // create 30 shares
+	k := byte(2)                  // require 2 of them to combine
 
 	shares, err := Split(n, k, []byte(secret)) // split into 30 shares
 	if err != nil {
@@ -16,10 +16,10 @@ func Example() {
 	}
 
 	// select a random subset of the total shares
-	subset := make(map[int][]byte, k)
+	subset := make(map[byte][]byte, k)
 	for x, y := range shares { // just iterate since maps are randomized
 		subset[x] = y
-		if len(subset) == k {
+		if len(subset) == int(k) {
 			break
 		}
 	}
